mocks/pyproject.toml
[project]
name = "pathology-api-mocks"
version = "0.1.0"
description = ""
authors = [
  {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">3.13,<4.0.0"
dependencies = [
  "pyjwt[crypto] (>=2.11.0,<3.0.0)",
  "aws-lambda-powertools (>=3.24.0,<4.0.0)",
  "pydantic (>=2.12.5,<3.0.0)",
  "boto3 (>=1.42.50,<2.0.0)"
]

[tool.poetry]
packages = [{include = "apim_mock", from = "src"}]

[tool.coverage.run]
relative_files = true
omit = ["*/tests/*", "*/features/*", "*/test_*.py"]

[tool.coverage.paths]
source = [
    "src/",
    "*/mocks/src/",
]
root = [
  ".",
  "*/mocks/"
]

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.html]
directory = "htmlcov"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
  "mypy (>=1.18.2,<2.0.0)",
  "pytest>=8.0.0",
  "pytest-bdd (>=8.1.0,<9.0.0)",
  "pytest-cov (>=7.0.0,<8.0.0)",
  "pytest-html (>=4.1.1,<5.0.0)",
  "requests>=2.31.0",
  "types-boto3 (>=1.42.51,<2.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
  "types-requests (>=2.32.4.20250913,<3.0.0.0)",
  "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
]

[tool.mypy]
strict = true
